refactor: Restart enum 리팩터링
Restart 이름을 더 알기 쉽도록 RestartOption로 변경.
RestartOption 객체를 받아오는 것이므로, validate가 아닌 create로 메서드 이름 변경.
정해진 option 이외(3)의 옵션을 검증하는 것은 RestartOption의 역할. 메서드 분리를 통해 validate함.
